[FIX] viewmodel분리 및 공통 interface분리(#31)
